Move Hdf4 code to internal, remove old.
