+++ cdm-test/src/test/java/ucar/nc2/internal/iosp/hdf4/TestH4eosRdAll.java
@@ -1,18 +1,20 @@
 /*
- * Copyright (c) 1998-2018 University Corporation for Atmospheric Research/Unidata
+ * Copyright (c) 1998-2020 John Caron and University Corporation for Atmospheric Research/Unidata
  * See LICENSE for license information.
  */
 
-package ucar.nc2.iosp.hdf4;
+package ucar.nc2.internal.iosp.hdf4;
 
-import org.junit.AfterClass;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import ucar.nc2.util.DebugFlagsImpl;
+import ucar.nc2.Group;
+import ucar.nc2.NetcdfFile;
+import ucar.nc2.NetcdfFiles;
+import ucar.nc2.Variable;
 import ucar.unidata.util.test.category.NeedsCdmUnitTest;
 import ucar.unidata.util.test.TestDir;
 import java.io.File;
@@ -21,29 +23,18 @@ import java.lang.invoke.MethodHandles;
 import java.util.ArrayList;
 import java.util.Collection;
 
-/**
- * Read all hdf4 files in cdmUnitTestDir + "formats/hdf4/"
- *
- * @author caron
- */
 @RunWith(Parameterized.class)
 @Category(NeedsCdmUnitTest.class)
-public class TestH4readAll {
+public class TestH4eosRdAll {
   private static final Logger logger = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
-  static private String testDir = TestDir.cdmUnitTestDir + "formats/hdf4/";
-
-  @AfterClass
-  static public void after() {
-    H4header.setDebugFlags(new DebugFlagsImpl("")); // make sure debug flags are off
-  }
 
   @Parameterized.Parameters(name = "{0}")
   public static Collection<Object[]> getTestParameters() throws IOException {
     Collection<Object[]> filenames = new ArrayList<>();
 
     try {
-      TestDir.actOnAllParameterized(testDir, new H4FileFilter(), filenames);
-      // TestDir.actOnAllParameterized("D:/hdf4/" , new H4FileFilter(), filenames);
+      H4EosFileFilter ff = new H4EosFileFilter();
+      TestDir.actOnAllParameterized(TestH4eos.testDir, ff, filenames);
     } catch (IOException e) {
       // JUnit *always* executes a test class's @Parameters method, even if it won't subsequently run the class's tests
       // due to an @Category exclusion. Therefore, we must not let it throw an exception, or else we'll get a build
@@ -58,21 +49,29 @@ public class TestH4readAll {
     return filenames;
   }
 
+  static class H4EosFileFilter implements java.io.FileFilter {
+    public boolean accept(File pathname) {
+      return pathname.getName().endsWith(".hdf") || pathname.getName().endsWith(".eos");
+    }
+  }
+
   String filename;
 
-  public TestH4readAll(String filename) {
+  public TestH4eosRdAll(String filename) {
     this.filename = filename;
   }
 
   @Test
-  public void readAll() throws IOException {
-    TestDir.readAll(filename);
-  }
+  public void testForStructMetadata() throws IOException {
+    System.out.printf("TestH4eosReadAll %s%n", filename);
+    try (NetcdfFile ncfile = NetcdfFiles.open(filename)) {
+      Group root = ncfile.getRootGroup();
+      Group g = root.findGroupLocal("HDFEOS INFORMATION");
+      if (g == null)
+        g = ncfile.getRootGroup();
 
-  static class H4FileFilter implements java.io.FileFilter {
-    public boolean accept(File pathname) {
-      return pathname.getName().endsWith(".hdf") || pathname.getName().endsWith(".eos")
-          || pathname.getName().endsWith(".h4");
+      Variable dset = g.findVariableLocal("StructMetadata.0");
+      assert (dset != null);
     }
   }
 }
