+++ uicdm/src/main/java/ucar/nc2/ui/menu/ModesMenu.java
@@ -12,7 +12,7 @@ import ucar.nc2.grib.GribData;
 import ucar.nc2.grib.grib1.tables.Grib1ParamTables;
 import ucar.nc2.internal.dataset.CoordSystemFactory;
 import ucar.nc2.internal.iosp.hdf5.H5iospNew;
-import ucar.nc2.iosp.hdf4.H4header;
+import ucar.nc2.internal.iosp.hdf4.H4header;
 import ucar.nc2.jni.netcdf.Nc4Iosp;
 import ucar.nc2.ui.ToolsUI;
 import ucar.ui.widget.BAMutil;
